+++ scripts/macros/2014/actions/genericActions.js
@@ -1,0 +1,7 @@
+import {genericActions as modernGenericActions} from '../../2024/actions/genericActions.js';
+export let genericActions = {
+    name: 'Generic Actions',
+    version: '1.3.115',
+    rules: 'legacy',
+    ...modernGenericActions
+};