+++ scripts/macros/2024/actions/fall.js
@@ -1,0 +1,6 @@
+import {fall as legacyFall} from '../../../legacyMacros.js';
+export let fall = {
+    ...legacyFall,
+    version: '1.3.115',
+    rules: 'modern'
+};