+++ scripts/macros/2024/actions/search.js
@@ -1,0 +1,5 @@
+export let search = {
+    name: 'Search',
+    version: '1.3.115',
+    rules: 'modern'
+};